refactor(models): keep DCAE SAAF residual blocks local
