+++ compressai/models/dcae.py
@@ -35,10 +35,6 @@ from compressai.latent_codecs import (
 )
 from compressai.layers.attn.dictionary import ConvolutionalGLU, Scale
 from compressai.layers.attn.swin import pad_to_window_multiple
-from compressai.layers.lic import (
-    ResidualBottleneckBlockWithStride,
-    ResidualBottleneckBlockWithUpsample,
-)
 from compressai.models._helpers.dictionary_context import (
     SharedDictionary,
     build_dictionary_mean_scale_head,
@@ -49,6 +45,7 @@ from compressai.models._helpers.slice_helpers import (
     make_entropy_transform,
 )
 from compressai.models.base import CompressionModel
+from compressai.models.sensetime import ResidualBottleneckBlock
 from compressai.models.utils import conv, deconv
 from compressai.registry import register_model
 
@@ -147,6 +144,40 @@ class _SideContextChannelGroupsLatentCodec(ChannelGroupsLatentCodec):
 # (Inlined from the upstream DCAE source rather than lifted to compressai/layers/
 # because they are not reused by other models in the PR series.)
 # ---------------------------------------------------------------------------
+
+
+class _ResidualBottleneckBlockWithStride(nn.Module):
+    """DCAE stride-2 residual-bottleneck downsampling block."""
+
+    def __init__(self, in_ch: int, out_ch: int) -> None:
+        super().__init__()
+        self.conv = conv(in_ch, out_ch, kernel_size=5, stride=2)
+        self.res1 = ResidualBottleneckBlock(out_ch, out_ch)
+        self.res2 = ResidualBottleneckBlock(out_ch, out_ch)
+        self.res3 = ResidualBottleneckBlock(out_ch, out_ch)
+
+    def forward(self, input_tensor: Tensor) -> Tensor:
+        output = self.conv(input_tensor)
+        output = self.res1(output)
+        output = self.res2(output)
+        return self.res3(output)
+
+
+class _ResidualBottleneckBlockWithUpsample(nn.Module):
+    """DCAE residual-bottleneck upsampling block."""
+
+    def __init__(self, in_ch: int, out_ch: int) -> None:
+        super().__init__()
+        self.res1 = ResidualBottleneckBlock(in_ch, in_ch)
+        self.res2 = ResidualBottleneckBlock(in_ch, in_ch)
+        self.res3 = ResidualBottleneckBlock(in_ch, in_ch)
+        self.conv = deconv(in_ch, out_ch, kernel_size=5, stride=2)
+
+    def forward(self, input_tensor: Tensor) -> Tensor:
+        output = self.res1(input_tensor)
+        output = self.res2(output)
+        output = self.res3(output)
+        return self.conv(output)
 
 
 class _WMSA(nn.Module):
@@ -455,7 +486,7 @@ class DCAE(CompressionModel):
 
         # ----- g_a / g_s -----
         self.g_a = nn.Sequential(
-            ResidualBottleneckBlockWithStride(input_image_channel, feature_dims[0]),
+            _ResidualBottleneckBlockWithStride(input_image_channel, feature_dims[0]),
             _SwinBlockWithConvMulti(
                 feature_dims[0],
                 feature_dims[0],
@@ -465,7 +496,7 @@ class DCAE(CompressionModel):
                 _ResScaleConvolutionGateBlock,
                 block_num=block_num[0],
             ),
-            ResidualBottleneckBlockWithStride(feature_dims[0], feature_dims[1]),
+            _ResidualBottleneckBlockWithStride(feature_dims[0], feature_dims[1]),
             _SwinBlockWithConvMulti(
                 feature_dims[1],
                 feature_dims[1],
@@ -475,7 +506,7 @@ class DCAE(CompressionModel):
                 _ResScaleConvolutionGateBlock,
                 block_num=block_num[1],
             ),
-            ResidualBottleneckBlockWithStride(feature_dims[1], feature_dims[2]),
+            _ResidualBottleneckBlockWithStride(feature_dims[1], feature_dims[2]),
             _SwinBlockWithConvMulti(
                 feature_dims[2],
                 feature_dims[2],
@@ -498,7 +529,7 @@ class DCAE(CompressionModel):
                 _ResScaleConvolutionGateBlock,
                 block_num=block_num[2],
             ),
-            ResidualBottleneckBlockWithUpsample(feature_dims[2], feature_dims[1]),
+            _ResidualBottleneckBlockWithUpsample(feature_dims[2], feature_dims[1]),
             _SwinBlockWithConvMulti(
                 feature_dims[1],
                 feature_dims[1],
@@ -508,7 +539,7 @@ class DCAE(CompressionModel):
                 _ResScaleConvolutionGateBlock,
                 block_num=block_num[1],
             ),
-            ResidualBottleneckBlockWithUpsample(feature_dims[1], feature_dims[0]),
+            _ResidualBottleneckBlockWithUpsample(feature_dims[1], feature_dims[0]),
             _SwinBlockWithConvMulti(
                 feature_dims[0],
                 feature_dims[0],
@@ -518,12 +549,12 @@ class DCAE(CompressionModel):
                 _ResScaleConvolutionGateBlock,
                 block_num=block_num[0],
             ),
-            ResidualBottleneckBlockWithUpsample(feature_dims[0], output_image_channel),
+            _ResidualBottleneckBlockWithUpsample(feature_dims[0], output_image_channel),
         )
 
         # ----- h_a / h_mean_s / h_scale_s -----
         h_a = nn.Sequential(
-            ResidualBottleneckBlockWithStride(M, N),
+            _ResidualBottleneckBlockWithStride(M, N),
             _SwinBlockWithConvMulti(
                 N,
                 N,
@@ -551,7 +582,7 @@ class DCAE(CompressionModel):
                 _ResScaleConvolutionGateBlock,
                 block_num=1,
             ),
-            ResidualBottleneckBlockWithUpsample(N, M),
+            _ResidualBottleneckBlockWithUpsample(N, M),
         )
         h_scale_s = nn.Sequential(
             deconv(hyper_channels, N, kernel_size=3, stride=2),
@@ -564,7 +595,7 @@ class DCAE(CompressionModel):
                 _ResScaleConvolutionGateBlock,
                 block_num=1,
             ),
-            ResidualBottleneckBlockWithUpsample(N, M),
+            _ResidualBottleneckBlockWithUpsample(N, M),
         )
 
         # ----- Shared dictionary -----
